Restrict ipaddress dependency to Python < 3.3
ifaddr depends on the ipaddress module. Before Python 3.3, ipaddress was
a separate module, but it’s now shipped with vanilla Python.

This means that from Python 3.3 onward, it’s not necessary to add
ipaddress as an install dependency anymore.

More than that, the explicit dependency is problematic, because users of
Python 3.3 and newer will get an error message. The reason is that
ipaddress won’t be found as a separate module anymore, even though it’s
implicitly available.

This patch restricts the ipaddress install dependency to Python versions
before 3.3. In this way, users of Python earlier than 3.3 will still be
required to get the ipaddress module backport, whereas users of newer
Python versions won’t be bothered with error messages.