fixes installation of ipaddress module
A user had reported that the install_requires= line of the setup causes
issues #14. This update changes it so that the testing of the python
version happens in the ifaddr package and not in distutuls. This is a
far better way to go about it due to the very large number of distutils
and setup tools versions available. not all versions support the same
syntax or the install_requires as well. but all versions do
support requires= with a string for the module name or an empty list
if there are no dependencies. This will remove the need to add special
lines to handle specifically the setup tools version. This package has a
pretty simple setup routine, so there is no need to have requirements for
specific setup tools versions.