Improve platorm-specific type safety
All these type: ignore directives were annoying me and potentially
hiding problems.

To get rid of them we have to switch to a couple of conditional
checks to a form recognized by Mypy, that is: sys.platform and direct
comparison or "startswith".

Additional assert statements inserted to tell Mypy to ignore the win32
module on non-Windows and vice versa – we'd get errors otherwise.