Split win32 code into IO layer and processing layer
The ctypes-based structures aren't type-checked too well which is
annoyin, so I figured we could first translate whatever the win32 API
gives us into a bunch of platform-independent objects that we could then
process.

The nice thing about it is we could them dump the data on any Windows
machine and unit-test the processing logic in isolation as needed.

Not all adapter data is fetched from the win32 API yet, we'll expand on
this.